references.strata: New script to test references
I'd have liked to done:

  let r: Ref Int = ref 1 in
  (
      print(!r);  // prints 1
      r <- 2;
      print(!r)   // prints 2
  )

but our parser currently gives the semicolons a higher associativity
than our parentheses. This leads to a parse of the following structure:

  seq
  |
  \ let r: Ref Int = ref 1 in ( print(!r)
  \ r <- 2
  \ print(!r) )

It splits the parentheses across difference expressions, which results
in two parse errors. You'll also notice that `r` is only defined in the
first expression within the seq, so the accesses to `r` within the
second and third seq expression fail their typechecks with an `unbound
variable` message.

This is a parsing bug and should be fixed.